backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
aiosqlite==0.19.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic==2.7.1
email-validator==2.1.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-multipart==0.0.9
